abstract out mergeCellBelow
